requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-huggingface>=0.1.0
anthropic>=0.40.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
rank_bm25>=0.2.2
python-dotenv>=1.0.0
